Stop using handlers
With the switch to tripleo_free in stable/train, we can no longer
reliably use handlers.  That being said, handlers don't make sense when
you are forcing them to run with a flush handlers anyway. So this change
takes the code we used in the handler and makes them explicit tasks in
the required locations where we used to do a flush handlers.  This
change should also improve the scope of some tasks to only occur when
necessary.

Change-Id: I1e31af24a4c1b9356a5971c6c1cb717d39d1e9b2
Closes-Bug: #1909750